requirements.txt
requests-oauthlib==1.3.0
python-dotenv==0.10.3
selenium==3.141.0
